Flatten coreboot payload modules
