+++ crabefi-coreboot/src/cbmem_console.rs
