+++ crabefi-coreboot/src/memory.rs
